requirements.txt
# https://pypi.org/project/flask
Flask==3.0.2

# https://pypi.org/project/gunicorn
gunicorn==21.2.0

# https://pypi.org/project/google-cloud-storage
google-cloud-storage==2.9.0

# Google Generative AI SDK (for Gemini API)
google-generativeai==0.7.2

# Google Cloud Secret Manager (for secure API key retrieval)
google-cloud-secret-manager==2.16.2

# https://pypi.org/project/setuptools
setuptools
